[Core] Slightly respecify the `operators` protocol
This is done to prepare for the fix with respect to ODR-violation.
Basically, the problem is that fixing ODR-violations and ensuring
a unique address for function objects requires `decltype(method)`
to be a reference. Hence, we can't inherit from `decltype(method)`
in the `operators::of` specializations anymore. This commit changes
`operators::of` to use a slightly different protocol.